Configure webpack to compile theme
When shipping NPM modules, they typically need to be either
pre-compiled or the user needs to add bundler config to process the
files. Gatsby lets us ship the bundler config *with* the theme, so
we never need a lib-side build step.  Since we dont pre-compile the
theme, this is how we let webpack know how to process files.